Changed values to double and made them round to integers

+++ ggc-core/src/ggc/Bargain.java
@@ -5,7 +5,7 @@ public class Bargain extends Notification{
   /** Serial number for serialization. */
   private static final long serialVersionUID = 202110262229L;
 
-  public Bargain(String productKey, int productPrice) {
+  public Bargain(String productKey, double productPrice) {
     super(productKey, productPrice);
   }
